+++ js/app/packages/core/component/Lightbox/index.ts
@@ -1,0 +1,2 @@
+export { Lightbox } from './Lightbox';
+export { LightboxToolbar } from './LightboxToolbar';
